Add support for additional windspeed units
The AWS-X station supports sending windspeeds in m/s, km/h and knots.
Since the transmitted packet does not contain which unit it uses, the
receiver (awsxd) must be configured with the same unit.

This patch adds a new command-line argument (-m) to select the unit,
which must correspond to the same unit configured at the station.

awsxd will internally store all windspeeds as m/s. For knots and km/h
transmissions, awsxd will convert the values before storage. For this
reason, this patch also alters the database structure for the windspeed
fields to FLOAT. See the table_structure.sql diff for details. For
migrating the old database content, you will want to divide all windspeed
fields (wind_avg, wind_max, wind_min and wind_daily_max) with 3.6 to
convert them from km/h to m/s.

+++ table_structure.sql
@@ -5,10 +5,10 @@ CREATE TABLE awsx (
   station varchar(255) NOT NULL,
   sms_counter int(11) NOT NULL,
   sample_interval int(11) NOT NULL,
-  wind_avg int(11) NOT NULL,
-  wind_min int(11) NOT NULL,
-  wind_max int(11) NOT NULL,
-  wind_daily_max int(11) NOT NULL,
+  wind_avg float NOT NULL,
+  wind_min float NOT NULL,
+  wind_max float NOT NULL,
+  wind_daily_max float NOT NULL,
   wind_dir int(11) NOT NULL,
   wind_stability int(11) NOT NULL,
   air_pressure int(11) NOT NULL,
@@ -29,4 +29,3 @@ CREATE TABLE awsx (
   battery_voltage float NOT NULL,
   PRIMARY KEY  (id)
 ) ENGINE=MyISAM  DEFAULT CHARSET=latin1;
-
